+++ lib/rig/event_stream/kafka_as_http.ex
@@ -16,7 +16,7 @@ defmodule Rig.EventStream.KafkaToHttp do
 
   # ---
 
-  def kafka_handler(message) do
+  def kafka_handler(message, _topic) do
     case CloudEvent.parse(message) do
       {:ok, %CloudEvent{} = cloud_event} ->
         Logger.debug(fn -> inspect(cloud_event.parsed) end)