+++ lib/rig_outbound_gateway/kinesis/java_client.ex
@@ -116,8 +116,10 @@ defmodule RigOutboundGateway.Kinesis.JavaClient do
 
   @spec java_client_callback(data :: [{atom(), String.t()}, ...]) :: :ok
   def java_client_callback(data) do
+    conf = config()
+
     data[:body]
     |> Poison.decode!()
-    |> KinesisToFilter.kinesis_handler()
+    |> KinesisToFilter.kinesis_handler(conf.kinesis_stream)
   end
 end